extract proto to proto module
