+++ authentication/src/main/kotlin/com/kohan/authentication/util/TokenGenerator.kt
@@ -1,6 +1,6 @@
 package com.kohan.authentication.util
 
-import com.kohan.shared.armeria.converter.request.result.AccessDeviceInfo
+import com.kohan.shared.collection.user.item.AccessDeviceInfo
 import com.kohan.shared.collection.user.item.TokenInfo
 import org.springframework.beans.factory.annotation.Value
 import org.springframework.stereotype.Component
