+++ message-rest/src/main/kotlin/com/kohan/message/rest/service/annotation/UserService.kt
@@ -1,0 +1,16 @@
+package com.kohan.message.rest.service.annotation
+
+import com.linecorp.armeria.server.annotation.Param
+import com.linecorp.armeria.server.annotation.Post
+import com.linecorp.armeria.server.annotation.ProducesJson
+import org.springframework.stereotype.Service
+
+@Service
+class UserService {
+    @Post("/user/nickname/{nickname}")
+    @ProducesJson
+    fun changeNickname(
+        @Param("nickname") nickname: String,
+    ) {
+    }
+}
