+++ shared-armeria/build.gradle.kts
@@ -1,9 +1,4 @@
-import com.google.protobuf.gradle.id
-import com.google.protobuf.gradle.protobuf
-
-
 plugins {
-    alias(libs.plugins.protobuf)
     alias(libs.plugins.spring.dependency.management)
 }
 
@@ -17,51 +12,13 @@ dependencyManagement {
 }
 
 dependencies {
-    protobuf(project(":proto"))
-
     api(libs.kafka)
 
     api(libs.armeria.kotlin)
     api(libs.armeria.grpc)
     api(libs.armeria.kafka)
 
-    api(libs.grpc.stub)
-    api(libs.grpc.protobuf)
-    api(libs.grpc.kotlin.stub)
-
-    api(libs.protobuf.kotlin)
-
-    implementation(libs.uap.java)
-
     runtimeOnly(libs.slf4j.simple)
-}
-
-protobuf {
-    protoc {
-        artifact = of(libs.artifact.protoc)
-    }
-
-    plugins {
-        id("grpc") {
-            artifact = of(libs.artifact.gen.grpc.java)
-        }
-        id("grpckt") {
-            artifact = of(libs.artifact.gen.grpc.kotlin) + ":jdk8@jar"
-        }
-    }
-
-    // generate code
-    generateProtoTasks {
-        all().forEach {
-            it.plugins {
-                id("grpc")
-                id("grpckt")
-            }
-            it.builtins {
-                id("kotlin")
-            }
-        }
-    }
 }
 
 fun of(dependency: Provider<MinimalExternalModuleDependency>): String {
