+++ shared-spring/src/main/kotlin/com/kohan/shared/spring/config/PushGrpcClient.kt
@@ -1,6 +1,6 @@
 package com.kohan.shared.spring.config
 
-import com.kohan.shared.armeria.push.v1.FCMTokenServiceGrpc.FCMTokenServiceBlockingStub
+import com.kohan.proto.push.v1.FCMTokenServiceGrpc.FCMTokenServiceBlockingStub
 import com.linecorp.armeria.client.grpc.GrpcClients
 import org.springframework.beans.factory.annotation.Value
 import org.springframework.context.annotation.Bean
