+++ shared-spring/src/main/kotlin/com/kohan/shared/spring/converter/request/parser/AccessInfoParser.kt
@@ -1,6 +1,6 @@
-package com.kohan.shared.armeria.converter.request.parser
+package com.kohan.shared.spring.converter.request.parser
 
-import com.kohan.shared.armeria.converter.request.result.AccessDeviceInfo
+import com.kohan.shared.collection.user.item.AccessDeviceInfo
 import ua_parser.Parser
 
 internal class AccessInfoParser {
