+++ shared-spring/src/test/kotlin/com/kohan/shared/spring/converter/request/parser/AccessInfoParserTest.kt
@@ -1,6 +1,6 @@
-package com.kohan.shared.armeria.converter.request.parser
+package com.kohan.shared.spring.converter.request.parser
 
-import com.kohan.shared.armeria.converter.request.result.AccessDeviceInfo
+import com.kohan.shared.collection.user.item.AccessDeviceInfo
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Test
